pyproject.toml
[project]
name = "livekit-info-bot"
version = "0.1.0"
requires-python = "==3.13.*"
dependencies = [
    "google-genai>=1.55.0",
    "livekit-agents[silero,turn-detector]~=1.2",
    "livekit-plugins-noise-cancellation~=0.2",
    "numpy>=2.3.5",
    "python-dotenv>=1.2.1",
    "scikit-learn>=1.8.0",
]
